Add Spell description in monster info dialog
Besides this:
- fix few spell messages in the game
- optimize ground type translation calls (no need to translate all in once)
- remove MAXU16 macro
- remove resource cost and state of spells. They were always 0. Relates to #3903
- add translations to some missing strings
- fix few compilation warnings
- fix one Polish sentence (missing capital letter)

+++ src/fheroes2/world/world.cpp
@@ -1007,8 +1007,8 @@ bool World::DiggingForUltimateArtifact( const fheroes2::Point & center )
     Maps::Tiles & tile = GetTiles( center.x, center.y );
 
     // puts hole sprite
-    int obj = 0;
-    u32 idx = 0;
+    uint8_t obj = 0;
+    uint32_t idx = 0;
 
     switch ( tile.GetGround() ) {
     case Maps::Ground::WASTELAND:
